Fixed graphiql enpoint issue by removing mapping in properties

+++ GraphQLSample/src/main/java/com/learning/GraphQLSample/repository/BooksRepository.java
@@ -4,5 +4,5 @@ import com.learning.GraphQLSample.entity.Books;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface BooksRepository extends JpaRepository<Books, Long> {
+public interface BooksRepository extends JpaRepository<Books, Integer> {
 }
